Fix bug in insert_meeting_creator

+++ trigger.sql
@@ -325,17 +325,19 @@ FOR EACH ROW EXECUTE FUNCTION check_resignation_updates();
 CREATE OR REPLACE FUNCTION insert_meeting_creator()
 RETURNS TRIGGER AS $$
 BEGIN
-    DELETE FROM Attends WHERE employee_id = OLD.creator_id AND floor = OLD.floor AND room = OLD.room AND date = OLD.date AND start_hour = OLD.start_hour;
-    IF (NOT EXISTS 
-        (SELECT * 
-        FROM Attends AS A 
-        WHERE A.employee_id = NEW.creator_id 
-            AND A.room = NEW.room 
-            AND A.floor = NEW.floor 
-            AND A.date = NEW.date 
-            AND A.start_hour = NEW.start_hour)
-    ) THEN
-        INSERT INTO Attends VALUES (NEW.creator_id, NEW.floor, NEW.room, NEW.date, NEW.start_hour);
+    IF (NEW.creator_id IS DISTINCT FROM OLD.creator_id) THEN
+        DELETE FROM Attends WHERE employee_id = OLD.creator_id AND floor = OLD.floor AND room = OLD.room AND date = OLD.date AND start_hour = OLD.start_hour;
+        IF (NOT EXISTS 
+            (SELECT * 
+            FROM Attends AS A 
+            WHERE A.employee_id = NEW.creator_id 
+                AND A.room = NEW.room 
+                AND A.floor = NEW.floor 
+                AND A.date = NEW.date 
+                AND A.start_hour = NEW.start_hour)
+        ) THEN
+            INSERT INTO Attends VALUES (NEW.creator_id, NEW.floor, NEW.room, NEW.date, NEW.start_hour);
+        END IF;
     END IF;
     RETURN NEW;
 END;
